split function into new file
